+++ modules/nodes/07-sg.tf
@@ -63,6 +63,6 @@ resource "aws_security_group_rule" "allow_kubelet_port_from_internal_subnets" {
   from_port         = 10250
   to_port           = 10250
   protocol          = "TCP"
-  cidr_blocks       = values(data.aws_subnet.target).*.cidr_block
+  cidr_blocks       = var.private_subnets_cidr
   security_group_id = aws_security_group.k8s_workers_node_sg.id
 }